# tools/WillowWSUP/WsupApi/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Willow.Support.Api",
    "version": "v1"
  },
  "paths": {
...
    "/Applications/Get": {
      "get": {
        "tags": [
          "Applications"
        ],
        "operationId": "Applications_Get",
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      }
    },
    "/Applications": {
      "post": {
        "tags": [
          "Applications"
        ],
        "operationId": "Applications_Create",
        "requestBody": {
          "x-name": "application",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Applications"
        ],
        "operationId": "Applications_Update",
        "requestBody": {
          "x-name": "application",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "operationId": "Applications_Delete",
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/BuildingConnectors": {
      "get": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildingConnector"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_Create",
        "requestBody": {
          "x-name": "buildingConnector",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildingConnector"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_Update",
        "requestBody": {
          "x-name": "buildingConnector",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildingConnector"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/BuildingConnectors/GetAllBuildingConnectorsByCustomerInstance/{customerInstanceId}": {
      "get": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_GetAll2",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildingConnector"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/BuildingConnectors/{customerInstanceId}/{buildingId}/{connectorId}": {
      "get": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_Get",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "buildingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "connectorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingConnector"
                }
              }
            }
          }
        }
      }
    },
    "/BuildingConnectors/{customerInstanceId}/{applicationId}": {
      "delete": {
        "tags": [
          "BuildingConnectors"
        ],
        "operationId": "BuildingConnectors_Delete",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "buildingId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "connectorId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 3
          },
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/Buildings": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Building"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_Create",
        "requestBody": {
          "x-name": "building",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Building"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_Update",
        "requestBody": {
          "x-name": "building",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Building"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Building"
                }
              }
            }
          }
        }
      }
    },
    "/Buildings/GetAllByCustomerInstanceId/{customerInstanceId}": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_GetAllByCustomerInstanceId",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Building"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Buildings/{buildingId}": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_Get",
        "parameters": [
          {
            "name": "buildingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Building"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Buildings"
        ],
        "operationId": "Buildings_Delete",
        "parameters": [
          {
            "name": "buildingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Application": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "teamId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "healthEndpointPath": {
            "type": "string",
            "nullable": true
          },
          "hasPublicEndpoint": {
            "type": "boolean"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Team"
              }
            ]
          },
          "applicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          }
        }
      },
      "Team": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "applications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Application"
            }
          }
        }
      },
      "ApplicationStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "applications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Application"
            }
          }
        }
      },
      "CustomerInstanceApplication": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstanceApplicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "application": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Application"
              }
            ]
          },
          "customerInstanceApplicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceApplicationStatus"
              }
            ]
          }
        }
      },
      "CustomerInstance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "customerId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "fullDomain": {
            "type": "string",
            "nullable": true
          },
          "dnsEnvSuffix": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "fullCustomerInstanceName": {
            "type": "string",
            "nullable": true
          },
          "deploymentPhase": {
            "type": "string",
            "nullable": true
          },
          "lifecycleState": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "azureDigitalTwinsInstance": {
            "type": "string",
            "nullable": true
          },
          "azureDataExplorerInstance": {
            "type": "string",
            "nullable": true
          },
          "stampId": {
            "type": "string",
            "format": "guid"
          },
          "customer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Customer"
              }
            ]
          },
          "customerInstanceStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceStatus"
              }
            ]
          },
          "stamp": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Stamp"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          },
          "buildings": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Building"
            }
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "Customer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "salesId": {
            "type": "string",
            "nullable": true
          },
          "customerStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "customerStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerStatus"
              }
            ]
          }
        }
      },
      "CustomerStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          }
        }
      },
      "CustomerInstanceStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "adxDatabaseUrl": {
            "type": "string",
            "nullable": true
          },
          "adtInstanceUrl": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          }
        }
      },
...
      "Building": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
      "BuildingConnector": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "buildingId": {
            "type": "string",
            "nullable": true
          },
          "connectorId": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "buildingConnectorStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "building": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Building"
              }
            ]
          },
          "connector": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Connector"
              }
            ]
          },
          "buildingConnectorStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/BuildingConnectorStatus"
              }
            ]
          }
        }
      },
      "Connector": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "connectorStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "connectorTypeId": {
            "type": "string",
            "nullable": true
          },
          "connectorStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorStatus"
              }
            ]
          },
          "connectorType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorType"
              }
            ]
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
      "ConnectorStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "ConnectorType": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "nullable": true
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "BuildingConnectorStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
      "CustomerInstanceApplicationStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          }
        }
      }
    }
  }
}